ref: Structured use basket inside basket provider
